+++ src/ast/rewriter/CMakeLists.txt
@@ -8,6 +8,7 @@ z3_add_component(rewriter
     bv_bounds.cpp
     bv_elim.cpp
     bv_rewriter.cpp
+    cached_var_subst.cpp
     datatype_rewriter.cpp
     der.cpp
     distribute_forall.cpp
