+++ src/ast/rewriter/cached_var_subst.h
