+++ src/smt/qi_queue.h
@@ -19,13 +19,13 @@ Revision History:
 #pragma once
 
 #include "ast/ast.h"
-#include "ast/cached_var_subst.h"
 #include "ast/quantifier_stat.h"
+#include "ast/rewriter/cached_var_subst.h"
+#include "parsers/util/cost_parser.h"
 #include "smt/smt_checker.h"
 #include "smt/smt_quantifier.h"
 #include "smt/fingerprints.h"
 #include "smt/params/qi_params.h"
-#include "parsers/util/cost_parser.h"
 #include "smt/cost_evaluator.h"
 #include "util/statistics.h"
